+++ match_model/reporting/manually_run_reports.py
@@ -6,12 +6,13 @@
 This script is used for testing and debugging the process of running and saving the summary report jupyter notebooks
 """
 
+from msilib.schema import File
 import os
 import shutil
 import pandas as pd
 
 # modify the file path to the name of the scenario for which you want to re-run the reporting
-run_folder = '../../MODEL_RUNS/test_grid_impact'
+run_folder = '../../MODEL_RUNS/pre_rfo_run'
 #run_folder = '../../../../Box/Supply/24x7 Time-Coincident Work/MODEL_RUNS/test'
 
 scenarios = os.listdir(f'{run_folder}/outputs')
@@ -30,6 +31,33 @@ for s in scenarios:
     os.system(f'jupyter nbconvert --to html --no-input --no-prompt {inputs_dir}/summary_report.ipynb --output-dir {outdir} --output summary_report_{s}')
     os.system(f'jupyter nbconvert --clear-output --inplace {inputs_dir}/summary_report.ipynb')
 
+#%%
+
+# run a single notebook
+
+from msilib.schema import File
+import os
+import shutil
+import pandas as pd
+
+# modify the file path to the name of the scenario for which you want to re-run the reporting
+run_folder = '../../MODEL_RUNS/shortlist_run_456'
+#run_folder = '../../../../Box/Supply/24x7 Time-Coincident Work/MODEL_RUNS/test'
+
+scenarios = os.listdir(f'{run_folder}/outputs')
+
+s = 'run_4_tc100_ABB_RA_NOpen'
+
+inputs_dir = f'{run_folder}/inputs/{s}'
+outdir = f'{run_folder}/outputs/{s}'
+print(s)
+
+shutil.copy('summary_report.ipynb', inputs_dir)
+#shutil.copy('summary_report_public.ipynb', inputs_dir)
+
+os.system(f'jupyter nbconvert --ExecutePreprocessor.kernel_name="python3" --to notebook --execute --inplace {inputs_dir}/summary_report.ipynb')
+os.system(f'jupyter nbconvert --to html --no-input --no-prompt {inputs_dir}/summary_report.ipynb --output-dir {outdir} --output summary_report_{s}')
+os.system(f'jupyter nbconvert --clear-output --inplace {inputs_dir}/summary_report.ipynb')
 
 
 #%%
@@ -40,30 +68,30 @@ import shutil
 import pandas as pd
 
 # modify the file path to the name of the scenario for which you want to re-run the reporting
-run_folder = '../../MODEL_RUNS/google validation'
+run_folder = '../../MODEL_RUNS/shortlist_run_456'
 
 scenarios = os.listdir(f'{run_folder}/outputs')
 
 i = 0
 for s in scenarios:
-    summary_file = f'{run_folder}/outputs/{s}/scenario_summary.csv'
-    #buildgen_file = f'{run_folder}/outputs/{s}/BuildGen.csv'
+        summary_file = f'{run_folder}/outputs/{s}/scenario_summary.csv'
+        #buildgen_file = f'{run_folder}/outputs/{s}/BuildGen.csv'
 
-    if i == 0:
-        df = pd.read_csv(summary_file, index_col=0)
-        """
-        df_build = pd.read_csv(buildgen_file, usecols=['GEN_BLD_YRS_1','BuildGen'])
-        df_build = df_build.rename(columns={'GEN_BLD_YRS_1':'generation_project','BuildGen':s})
-        """
-        i += 1
-    else:
-        df2 = pd.read_csv(summary_file, index_col=0)
-        df = df.merge(df2, how='outer', left_index=True, right_index=True, sort=False)
-        """
-        df_build2 = pd.read_csv(buildgen_file, usecols=['GEN_BLD_YRS_1','BuildGen'])
-        df_build2 = df_build2.rename(columns={'GEN_BLD_YRS_1':'generation_project','BuildGen':s})
-        df_build = df_build.merge(df_build2, how='outer', on='generation_project')
-        """
+        if i == 0:
+            try:
+                df = pd.read_csv(summary_file, index_col=0)
+            except FileNotFoundError:
+                df = pd.DataFrame(data=['output not found'], columns=[s], index=['Scenario Name'])
+            i += 1
+        else:
+            try:
+                df2 = pd.read_csv(summary_file, index_col=0)
+            except FileNotFoundError:
+                df2 = pd.DataFrame(data=['output not found'], columns=[s], index=['Scenario Name'])
+            df = df.merge(df2, how='outer', left_index=True, right_index=True, sort=False)
+
+   
+
 
 #df = df.reset_index()
 #df['index'] = [i.split('~')[1] for i in df['index']]